+++ server/lib/tests/index.js
@@ -40,3 +40,4 @@ import './subtaskSettings.tests';
 import './subtaskCreation.tests';
 import './deleteWebhookActivity.tests';
 import './webhookNonBlocking.tests';
+import './apiResponseHelpers.tests';